datetimeをちゃんと使う

+++ db/migrate/001_create_table_miruene_records.rb
@@ -1,8 +1,8 @@
 class CreateTableMirueneRecords < ActiveRecord::Migration
 	def change
 		create_table :miruene_records do |t|
-			t.timestamp :datetime
-			t.float     :energy
+			t.timestamp  :datetime
+			t.integer    :energy
 			t.timestamps
 		end
 	end
